# Conflicts:
#	schema/schema.sql
#	src/main/java/io/hops/metadata/ndb/ClusterjConnector.java
#	src/main/java/io/hops/metadata/ndb/NdbStorageFactory.java

+++ src/main/java/io/hops/metadata/ndb/dalimpl/yarn/YarnProjectsDailyCostClusterJ.java
@@ -79,9 +79,9 @@ public class YarnProjectsDailyCostClusterJ implements
     void setDay(long day);
 
     @Column(name = CREDITS_USED)
-    int getCreditUsed();
+    float getCreditUsed();
 
-    void setCreditUsed(int credit);
+    void setCreditUsed(float credit);
 
   }
